singledog/requirements.txt

Flask-SQLAlchemy==2.1
Flask-Script==2.0.5
Flask-Login==0.3.2
Flask-WTF
PyMySQL==0.6.7
